Develop

+++ includes/class-settings.php
@@ -75,7 +75,7 @@ class Settings {
 			if ( $attribs['value'] ) {
 				$settings[ $field ] = $attribs['value'];
 			} elseif ( $attribs['default'] ) {
-				if ( $settings[ $field ] === null || $settings[ $field ] === '' || ( $settings[ $field ] === false && $attribs['type'] !== 'boolean' ) ) {
+				if ( null === $settings[ $field ] || '' === $settings[ $field ] || ( false === $settings[ $field ] && 'boolean' !== $attribs['type'] ) ) {
 					$settings[ $field ] = $attribs['default'];
 				}
 			}
@@ -94,7 +94,12 @@ class Settings {
 	 */
 	public static function save() {
 		$settings = apply_filters( 'murmurations_aggregator_save_settings', self::$settings );
-		return update_option( 'murmurations_aggregator_settings', $settings );
+		$existing = get_option( 'murmurations_aggregator_settings' );
+		if ( $existing !== $settings ) {
+			return update_option( 'murmurations_aggregator_settings', $settings );
+		} else {
+			return true;
+		}
 	}
 
 	/**
@@ -103,6 +108,7 @@ class Settings {
 	 * @return string Schema JSON
 	 */
 	private static function get_schema_json() {
+		// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents
 		return file_get_contents( MURMAG_ROOT_PATH . 'admin_fields_jschema.json' );
 	}
